+++ appbase/src/main/java/com/example/appbase/bean/CartListBean.java
@@ -126,7 +126,7 @@ public class CartListBean  extends BaseResult {
             private String unique;
             private double price;
             private String image;
-            private int commodityNum;
+            private double commodityNum;
             /**
              * 0上架；1下架
              */
@@ -230,11 +230,11 @@ public class CartListBean  extends BaseResult {
                 this.image = image;
             }
 
-            public int getCommodityNum() {
+            public double getCommodityNum() {
                 return commodityNum;
             }
 
-            public void setCommodityNum(int commodityNum) {
+            public void setCommodityNum(double commodityNum) {
                 this.commodityNum = commodityNum;
             }
