+++ appbase/src/main/java/com/example/appbase/bean/SellOrderDetailBean.java
@@ -686,7 +686,7 @@ public class SellOrderDetailBean extends BaseResult implements Parcelable {
         private String shopName;
         private double prices;
         private double price;
-        private int commodityNum;
+        private double commodityNum;
         private String cartInfo;
         private String unique;
 
@@ -780,11 +780,11 @@ public class SellOrderDetailBean extends BaseResult implements Parcelable {
             this.prices = prices;
         }
 
-        public int getCommodityNum() {
+        public double getCommodityNum() {
             return commodityNum;
         }
 
-        public void setCommodityNum(int commodityNum) {
+        public void setCommodityNum(double commodityNum) {
             this.commodityNum = commodityNum;
         }
 
@@ -824,7 +824,7 @@ public class SellOrderDetailBean extends BaseResult implements Parcelable {
             dest.writeString(this.shopName);
             dest.writeDouble(this.prices);
             dest.writeDouble(this.price);
-            dest.writeInt(this.commodityNum);
+            dest.writeDouble(this.commodityNum);
             dest.writeString(this.cartInfo);
             dest.writeString(this.unique);
             dest.writeInt(this.orderStatus);
@@ -841,7 +841,7 @@ public class SellOrderDetailBean extends BaseResult implements Parcelable {
             this.shopName = in.readString();
             this.prices = in.readDouble();
             this.price = in.readDouble();
-            this.commodityNum = in.readInt();
+            this.commodityNum = in.readDouble();
             this.cartInfo = in.readString();
             this.unique = in.readString();
             this.orderStatus = in.readInt();
